commit for multi-jump

+++ robot_control/base_controllers/utils/ros_publish.py
@@ -9,13 +9,14 @@ from visualization_msgs.msg import MarkerArray
 from geometry_msgs.msg import Point
 from geometry_msgs.msg import Polygon
 import numpy as np
+import pinocchio as pin
 
 import roslaunch
 import os
 import rospkg
 
 class RosPub():
-    def __init__(self, robot_name="solo", only_visual = False, visual_frame = "world"):
+    def __init__(self, robot_name="solo", only_visual = False, visual_frame = "world", markers_time_to_live = 0.):
 
         print("Starting ros pub---------------------------------------------------------------")
         if (not only_visual):                           
@@ -35,9 +36,11 @@ class RosPub():
         #init ros node to publish joint states and vis topics
         ros.init_node('sub_pub_node_python', anonymous=False, log_level=ros.FATAL)
 
+        self.markers_time_to_live = markers_time_to_live
         self.marker_pub = ros.Publisher('/vis' , MarkerArray, queue_size=1)
         self.arrow_pub = ros.Publisher('/arrow', MarkerArray, queue_size=1)
         self.polygon_pub = ros.Publisher('/support_polygon', MarkerArray, queue_size=1)
+        self.mesh_pub = ros.Publisher('/mesh', MarkerArray, queue_size=1)
         self.marker_fixed_pub = ros.Publisher('/point_fixed', MarkerArray, queue_size=1)
         self.markerArray = MarkerArray()
         self.markerArray.markers = []
@@ -47,10 +50,13 @@ class RosPub():
         self.markerArray_arrows.markers = []
         self.markerArrayFixed = MarkerArray()
         self.markerArrayFixed.markers = []
+        self.markerArray_mesh = MarkerArray()
+        self.markerArray_mesh.markers = []
         self.id = 0
         self.id_arrow = 0
         self.id_polygon = 0
         self.id_fixed = 0
+        self.id_mesh = 0
 
         self.fixedBaseRobot = False
         self.visual_frame = visual_frame
@@ -81,7 +87,7 @@ class RosPub():
         self.joint_pub.publish(msg)
         self.publishVisual()                                   
  
-    def publishVisual(self):                                
+    def publishVisual(self, delete_markers = True):
         #publish also the markers if any
         if len(self.markerArray.markers)>0:
             self.marker_pub.publish(self.markerArray)
@@ -104,12 +110,17 @@ class RosPub():
             self.markerArrayFixed.markers.clear()
             self.id_fixed = 0
 
+        if len(self.markerArray_mesh.markers) > 0:
+            self.mesh_pub.publish(self.markerArray_mesh)
+            # reset the marker array making it ready for another round
+            self.markerArray_mesh.markers.clear()
+            self.id_mesh = 0
 
 
+        if delete_markers:
+            self.delete_all_markers()
 
-        self.delete_all_markers()
-
-    def add_marker(self, pos, radius = 0.1, color = "red"):
+    def add_marker(self, pos, radius = 0.1, color = "red", alpha = 0.5):
         marker = Marker()
         marker.header.frame_id = self.visual_frame
         marker.type = marker.SPHERE
@@ -118,18 +129,32 @@ class RosPub():
         marker.scale.y = radius
         marker.scale.z = radius
         marker.color.a = 0.5
-        if (color == "red"):
-           marker.color.r = 1.0
-           marker.color.g = 0.0
-           marker.color.b = 0.0
-        if (color == "blue"):
-           marker.color.r = 0.0
-           marker.color.g = 0.0
-           marker.color.b = 1.0
-        if (color == "green"):
-           marker.color.r = 0.0
-           marker.color.g = 1.0
-           marker.color.b = 0.0
+        if isinstance(color, np.ndarray):
+            marker.color.r = color[0]
+            marker.color.g = color[1]
+            marker.color.b = color[2]
+        else:
+            if (color == "red"):
+               marker.color.r = 1.0
+               marker.color.g = 0.0
+               marker.color.b = 0.0
+            if (color == "blue"):
+               marker.color.r = 0.0
+               marker.color.g = 0.0
+               marker.color.b = 1.0
+            if (color == "green"):
+               marker.color.r = 0.0
+               marker.color.g = 1.0
+               marker.color.b = 0.0
+            if (color == "purple"):
+                marker.color.r = 0.7
+                marker.color.g = 0.0
+                marker.color.b = 1.0
+            if (color == "white"):
+                marker.color.r = 1.
+                marker.color.g = 1.
+                marker.color.b = 1.
+        marker.color.a = alpha
         marker.pose.orientation.x = 0.
         marker.pose.orientation.y = 0.
         marker.pose.orientation.z = 0.
@@ -137,8 +162,59 @@ class RosPub():
         marker.pose.position.x = pos[0]
         marker.pose.position.y = pos[1]
         marker.pose.position.z = pos[2]
-        marker.lifetime = ros.Duration(0.0)
+        marker.lifetime = ros.Duration(self.markers_time_to_live)
        
+        marker.id = self.id
+        self.id += 1
+        self.markerArray.markers.append(marker)
+
+    def add_plane(self, pos=np.array([0,0,0]), orient = np.array([0,0,0]),  color="red", alpha=0.5):
+        marker = Marker()
+        marker.header.frame_id = self.visual_frame
+        marker.type = marker.CUBE
+        marker.action = marker.ADD
+        marker.scale.x = 100
+        marker.scale.y = 100
+        marker.scale.z = 0.1
+        marker.color.a = 0.5
+        if isinstance(color, np.ndarray):
+            marker.color.r = color[0]
+            marker.color.g = color[1]
+            marker.color.b = color[2]
+        else:
+            if (color == "red"):
+                marker.color.r = 1.0
+                marker.color.g = 0.0
+                marker.color.b = 0.0
+            if (color == "blue"):
+                marker.color.r = 0.0
+                marker.color.g = 0.0
+                marker.color.b = 1.0
+            if (color == "green"):
+                marker.color.r = 0.0
+                marker.color.g = 1.0
+                marker.color.b = 0.0
+            if (color == "purple"):
+                marker.color.r = 0.7
+                marker.color.g = 0.0
+                marker.color.b = 1.0
+            if (color == "white"):
+                marker.color.r = 1.
+                marker.color.g = 1.
+                marker.color.b = 1.
+        marker.color.a = alpha
+
+
+        quaternion = pin.Quaternion(pin.rpy.rpyToMatrix(orient))
+        marker.pose.orientation.x = quaternion.x
+        marker.pose.orientation.y = quaternion.y
+        marker.pose.orientation.z = quaternion.z
+        marker.pose.orientation.w = quaternion.w
+        marker.pose.position.x = pos[0]
+        marker.pose.position.y = pos[1]
+        marker.pose.position.z = pos[2]
+        marker.lifetime = ros.Duration(self.markers_time_to_live)
+
         marker.id = self.id
         self.id += 1
         self.markerArray.markers.append(marker)
@@ -152,18 +228,24 @@ class RosPub():
         marker.scale.y = radius
         marker.scale.z = radius
         marker.color.a = 0.5
-        if (color == "red"):
-            marker.color.r = 1.0
-            marker.color.g = 0.0
-            marker.color.b = 0.0
-        if (color == "blue"):
-            marker.color.r = 0.0
-            marker.color.g = 0.0
-            marker.color.b = 1.0
-        if (color == "green"):
-            marker.color.r = 0.0
-            marker.color.g = 1.0
-            marker.color.b = 0.0
+        if isinstance(color, np.ndarray):
+            marker.color.r = color[0]
+            marker.color.g = color[1]
+            marker.color.b = color[2]
+        else:
+            if (color == "red"):
+                marker.color.r = 1.0
+                marker.color.g = 0.0
+                marker.color.b = 0.0
+            if (color == "blue"):
+                marker.color.r = 0.0
+                marker.color.g = 0.0
+                marker.color.b = 1.0
+            if (color == "green"):
+                marker.color.r = 0.0
+                marker.color.g = 1.0
+                marker.color.b = 0.0
+
         marker.pose.orientation.x = 0.
         marker.pose.orientation.y = 0.
         marker.pose.orientation.z = 0.
@@ -171,30 +253,40 @@ class RosPub():
         marker.pose.position.x = pos[0]
         marker.pose.position.y = pos[1]
         marker.pose.position.z = pos[2]
-        marker.lifetime = ros.Duration(0.0)
+        marker.lifetime = ros.Duration(self.markers_time_to_live)
 
         marker.id = self.id_fixed
         self.id_fixed += 1
         self.markerArrayFixed.markers.append(marker)
                             
-    def add_arrow(self, start, vector, color = "green", scale = 1.):
+    def add_arrow(self, start, vector, color = "green", scale = 1., alpha = 1.0):
        marker = Marker()
-       if (color == "green"):
-           marker.color.r = 0.0
-           marker.color.g = 1.0
-           marker.color.b = 0.0
-       if (color == "blue"):
-           marker.color.r = 0.0
-           marker.color.g = 0.0
-           marker.color.b = 1.0
-       if (color == "red"):
-           marker.color.r = 1.0
-           marker.color.g = 0.0
-           marker.color.b = 0.0
-       if (color == "black"):
-           marker.color.r = 1.0
-           marker.color.g = 1.0
-           marker.color.b = 1.0
+
+       if isinstance(color, np.ndarray):
+           marker.color.r = color[0]
+           marker.color.g = color[1]
+           marker.color.b = color[2]
+       else:
+           if (color == "green"):
+               marker.color.r = 0.0
+               marker.color.g = 1.0
+               marker.color.b = 0.0
+           if (color == "blue"):
+               marker.color.r = 0.0
+               marker.color.g = 0.0
+               marker.color.b = 1.0
+           if (color == "red"):
+               marker.color.r = 1.0
+               marker.color.g = 0.0
+               marker.color.b = 0.0
+           if (color == "black"):
+               marker.color.r = 0.0
+               marker.color.g = 0.0
+               marker.color.b = 0.0
+           if (color == "white"):
+               marker.color.r = 1.0
+               marker.color.g = 1.0
+               marker.color.b = 1.0
 
        marker.header.frame_id = self.visual_frame
        marker.type = marker.ARROW
@@ -204,8 +296,8 @@ class RosPub():
        marker.scale.x = 0.02*scale
        marker.scale.y = 0.04*scale
        marker.scale.z = 0.02*scale
-       marker.color.a = 1.0
-       marker.lifetime = ros.Duration(0.0)
+       marker.color.a = alpha
+       marker.lifetime = ros.Duration(self.markers_time_to_live)
        marker.pose.orientation.x = 0.
        marker.pose.orientation.y = 0.
        marker.pose.orientation.z = 0.
@@ -214,24 +306,81 @@ class RosPub():
        self.id_arrow += 1
        self.markerArray_arrows.markers.append(marker)
 
+    def add_mesh(self, package=None, mesh_path="", position = np.zeros(3), color = "green", alpha = 1.):
+        marker = Marker()
+        if color is None:
+            marker.color.r = 1.0
+            marker.color.g = 1.0
+            marker.color.b = 1.0
+            marker.color.a = 1.0
+        else:
+            if isinstance(color, np.ndarray):
+                marker.color.r = color[0]
+                marker.color.g = color[1]
+                marker.color.b = color[2]
+            else:
+                if (color == "green"):
+                    marker.color.r = 0.0
+                    marker.color.g = 1.0
+                    marker.color.b = 0.0
+                if (color == "blue"):
+                    marker.color.r = 0.0
+                    marker.color.g = 0.0
+                    marker.color.b = 1.0
+                if (color == "red"):
+                    marker.color.r = 1.0
+                    marker.color.g = 0.0
+                    marker.color.b = 0.0
+            marker.color.a = alpha
+
+        marker.header.frame_id = self.visual_frame
+        marker.type = marker.MESH_RESOURCE
+        if package is not None:
+            marker.mesh_resource = "package://" + package + mesh_path
+        else:
+            marker.mesh_resource =  "file://"+mesh_path
+        marker.mesh_use_embedded_materials = True  # Need this to use textures for mesh
+        marker.action = marker.ADD
+        marker.scale.x = 1
+        marker.scale.y = 1
+        marker.scale.z = 1
+        marker.lifetime = ros.Duration(self.markers_time_to_live)
+        marker.pose.position.x = position[0]
+        marker.pose.position.y = position[1]
+        marker.pose.position.z = position[2]
+        marker.pose.orientation.x = 0.
+        marker.pose.orientation.y = 0.
+        marker.pose.orientation.z = 0.
+        marker.pose.orientation.w = 1.
+        marker.id = self.id_mesh
+        self.id_mesh += 1
+        self.markerArray_mesh.markers.append(marker)
+
+
+
     def add_polygon(self, points, color = "green", scale = 1., visual_frame = 'world'):
         # list of points to connect:
         # a line connect points[0] - points[1]
         # a line connect points[1] - points[2]
         # ...
         marker = Marker()
-        if (color == "green"):
-            marker.color.r = 0.0
-            marker.color.g = 1.0
-            marker.color.b = 0.0
-        if (color == "blue"):
-            marker.color.r = 0.0
-            marker.color.g = 0.0
-            marker.color.b = 1.0
-        if (color == "red"):
-            marker.color.r = 1.0
-            marker.color.g = 0.0
-            marker.color.b = 0.0
+        if isinstance(color, np.ndarray):
+            marker.color.r = color[0]
+            marker.color.g = color[1]
+            marker.color.b = color[2]
+        else:
+            if (color == "green"):
+                marker.color.r = 0.0
+                marker.color.g = 1.0
+                marker.color.b = 0.0
+            if (color == "blue"):
+                marker.color.r = 0.0
+                marker.color.g = 0.0
+                marker.color.b = 1.0
+            if (color == "red"):
+                marker.color.r = 1.0
+                marker.color.g = 0.0
+                marker.color.b = 0.0
         marker.color.a = 1.0
         if visual_frame is None:
             marker.header.frame_id = self.visual_frame
@@ -244,7 +393,7 @@ class RosPub():
         marker.scale.x = 0.01 * scale
         marker.scale.y = 0.01 * scale
         marker.scale.z = 0.01 * scale
-        marker.lifetime = ros.Duration(0.0)
+        marker.lifetime = ros.Duration(self.markers_time_to_live)
         marker.pose.orientation.x = 0.
         marker.pose.orientation.y = 0.
         marker.pose.orientation.z = 0.
@@ -260,21 +409,29 @@ class RosPub():
         marker.id = 0
         marker.action = Marker.DELETEALL
         marker_array_msg.markers.append(marker)
+        self.marker_pub.publish(marker_array_msg)
         self.arrow_pub.publish(marker_array_msg)
+        self.polygon_pub.publish(marker_array_msg)
+        self.mesh_pub.publish(marker_array_msg)
 
     def add_cone(self,  origin, normal, friction_coeff, height=0.05, color = "green"):
 
        radius = friction_coeff* height
        tail_end = origin + normal*height; 
        marker = Marker()
-       if (color == "green"):                    
-           marker.color.r = 0.0
-           marker.color.g = 1.0
-           marker.color.b = 0.0                    
-       if (color == "blue"):                    
-           marker.color.r = 0.0
-           marker.color.g = 0.0
-           marker.color.b = 1.0                                                
+       if isinstance(color, np.ndarray):
+           marker.color.r = color[0]
+           marker.color.g = color[1]
+           marker.color.b = color[2]
+       else:
+           if (color == "green"):
+               marker.color.r = 0.0
+               marker.color.g = 1.0
+               marker.color.b = 0.0
+           if (color == "blue"):
+               marker.color.r = 0.0
+               marker.color.g = 0.0
+               marker.color.b = 1.0
 
        marker.header.frame_id = self.visual_frame
        marker.type = marker.ARROW
